+++ tests/testthat/test-vars-select.R
@@ -66,5 +66,24 @@ test_that("can supply empty inputs", {
 test_that("unknown variables errors are ignored if `.strict` is FALSE", {
   expect_identical(vars_select(letters, `_foo`, .strict = FALSE), set_names(chr()))
   expect_identical(vars_select(letters, a, `_foo`, .strict = FALSE), c(a = "a"))
+  expect_identical(vars_select(letters, a, "_foo", .strict = FALSE), c(a = "a"))
+
+  expect_identical(vars_select(letters, a, -`_foo`, .strict = FALSE), c(a = "a"))
+  expect_identical(vars_select(letters, a, -"`_foo`", .strict = FALSE), c(a = "a"))
+
   expect_identical(vars_select(letters, c(a, `_foo`, c), .strict = FALSE), c(a = "a", c = "c"))
+  expect_identical(vars_select(letters, c(a, "_foo", c), .strict = FALSE), c(a = "a", c = "c"))
+})
+
+test_that("`:` handles strings", {
+  expect_identical(vars_select(letters, "b":"d"), vars_select(letters, b:d))
+  expect_error(vars_select(letters, "b":"Z"), "Unknown column `Z`")
+})
+
+test_that("`-` handles strings", {
+  expect_identical(vars_select(letters, -"c"), vars_select(letters, -c))
+})
+
+test_that("`-` handles positions", {
+  expect_identical(vars_select(letters, 10 - 7), vars_select(letters, 3))
 })