+++ NEWS.md
@@ -55,8 +55,8 @@ vars_select(names(mtcars), .data$cyl : .data$drat)
 
 ## New features
 
-* `:` now handles strings as well. This makes it easy to unquote a
-  column name: `(!! name) : last_var()`.
+* `:` and `-` now handle strings as well. This makes it easy to
+  unquote a column name: `(!! name) : last_var()` or `-(!! name)`.
 
 * `vars_select()` gains a `.strict` argument similar to
   `rename_vars()`.  If set to `FALSE`, errors about unknown variables